+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to `wind` will be documented in this file
 
+## 1.1.0 - 2022-03-29
+
+- add support for laravel 9
+- use filament forms as the frontend
+- Improve the admin and the forms
+
 ## 1.0.0 - 2022-02-10
 
 - refactor Categories to Departments